modified ij.jar to see the Nifti Header. Niftis are better supported overall. BugFix on reading the Thickness. Open in External running a command depending on the current filepath from now on.

+++ src/gui/volumetab/SphereRoi.java
@@ -29,7 +29,7 @@ public class SphereRoi extends Roi implements Roi3D {
 	public void draw(Volume vol, BufferedImage bigimg, int slice, double scaling) {
 		if (this.getProperty("unit").equals("mm")) {
 			BufferedImage orig = vol.getSlice(0).getData().getBufferedImage();
-			double thickness = Integer.parseInt(vol.getSlice(0).getAttribute(
+			double thickness = Double.parseDouble(vol.getSlice(0).getAttribute(
 					KeyMap.KEY_SLICE_THICKNESS));
 			Rectangle rec = this.getBounds();
 			Roi3D roi3 = (Roi3D) this;
